修改success error 为api_return

+++ app/api/controller/Cart.php
@@ -36,7 +36,8 @@ class Cart extends ApiController
         $get = $this->request->get();
         if (empty($user)) {
             Sessions('member_id', null);
-            $this->error('用户信息异常请重新登录');
+//            $this->error('用户信息异常请重新登录');
+            return api_return(0, '用户信息异常请重新登录');
         }
         $carts = $user->carts()
             ->order('id', 'desc')
@@ -51,7 +52,8 @@ class Cart extends ApiController
             $item->goods = $goods;
             $item->is_valid = $goods->status === 1 ? true : false;
         }
-        $this->success('请求成功', ['carts_count' => $carts_count, 'list_count' => $list_count, 'carts' => $carts]);
+//        $this->success('请求成功', ['carts_count' => $carts_count, 'list_count' => $list_count, 'carts' => $carts]);
+        return api_return(1, '请求成功', ['carts_count' => $carts_count, 'list_count' => $list_count, 'carts' => $carts]);
     }
 
 
@@ -62,7 +64,8 @@ class Cart extends ApiController
         $user = Member::where('id', $user_id)->find();
         if (empty($user)) {
             Sessions('member_id', null);
-            $this->error('用户信息异常请重新登录');
+//            $this->error('用户信息异常请重新登录');
+            return api_return(0, '用户信息异常请重新登录');
         }
         $post = $this->request->post();
         $goods_id = isset($post['goods_id']) ? $post['goods_id'] : null;
@@ -75,17 +78,22 @@ class Cart extends ApiController
         $validate_result = $this->validate($goods_data, $message);
 
         if ($validate_result !== true) {
-            $this->error('请求参数有误请稍后重试');
+//            $this->error('请求参数有误请稍后重试');
+            return api_return(0, '请求参数有误请稍后重试');
         }
         $goods_obj = Goods::where('id', $goods_id)->where('status', 1)->find();
-        empty($goods_obj) && $this->error('商品不存在或已下架');
+//        empty($goods_obj) && $this->error('商品不存在或已下架');
+        if(empty($goods_obj)){
+            return api_return(0, '商品不存在或已下架');
+        }
         // 已存在购物车商品对象
         $old_objs = CartModel::where('goods_id', $goods_id)->where('uid', $user_id)->select();
 
         $goods_description = isset($post['description']) && is_array(json_decode($post['description'], true)) ? json_decode($post['description'], true) : null;
         $option = $goods_obj->isset_description($goods_description);
         if ($option === false) {
-            $this->error('商品规格选择错误');
+//            $this->error('商品规格选择错误');
+            return api_return(0, '商品规格选择错误');
         }
 
         // 判断之前购物车是否存在相同规格的商品
@@ -116,9 +124,11 @@ class Cart extends ApiController
                     throw new Exception('加入购物车失败了');
                 }
             } catch (Exception $e) {
-                $this->error('加入购物车失败,请稍后重试');
+//                $this->error('加入购物车失败,请稍后重试');
+                return api_return(0, '加入购物车失败,请稍后重试');
             }
-            $this->success('加入购物车成功');
+//            $this->success('加入购物车成功');
+            return api_return(1, '加入购物车成功');
         }
         if ($goods_num > 200) {
             $goods_num = 200;
@@ -141,9 +151,11 @@ class Cart extends ApiController
                 throw new Exception('加入购物车失败了');
             }
         } catch (Exception $e) {
-            $this->error('加入购物车失败,请稍后重试');
+//            $this->error('加入购物车失败,请稍后重试');
+            return api_return(0, '加入购物车失败,请稍后重试');
         }
-        $this->success('加入购物车成功');
+//        $this->success('加入购物车成功');
+        return api_return(1, '加入购物车成功');
     }
 
 
@@ -154,13 +166,17 @@ class Cart extends ApiController
         $user = Member::where('id', $user_id)->find();
         if (empty($user)) {
             Sessions('member_id', null);
-            $this->error('用户信息异常请重新登录');
+//            $this->error('用户信息异常请重新登录');
+            return api_return(0, '用户信息异常请重新登录');
         }
         $post = $this->request->post();
         $cart_goods_id = isset($post['cart_goods_id']) ? $post['cart_goods_id'] : null;
         $cart_goods_obj = CartModel::where('id', $cart_goods_id)->find();
         $goods = $cart_goods_obj->goods;
-        empty($cart_goods_obj) && $this->error('购物车商品已删除或不存在');
+//        empty($cart_goods_obj) && $this->error('购物车商品已删除或不存在');
+        if(empty($cart_goods_obj)){
+            return api_return(0, '购物车商品已删除或不存在');
+        }
         $goods_num = isset($post['goods_num']) ? $post['goods_num'] : null;
         $goods_data = ['cart_goods_id' => $cart_goods_id, 'goods_num' => $goods_num];
         $message = [
@@ -169,7 +185,8 @@ class Cart extends ApiController
         ];
         $validate_result = $this->validate($goods_data, $message);
         if ($validate_result !== true) {
-            $this->error('请求参数有误请稍后重试');
+//            $this->error('请求参数有误请稍后重试');
+            return api_return(0, '请求参数有误请稍后重试');
         }
         if ($goods_num > 200) {
             $goods_num = 200;
@@ -181,9 +198,11 @@ class Cart extends ApiController
                 throw new Exception('错误请稍后重试');
             }
         } catch (Exception $e) {
-            $this->error('修改失败请稍后重试');
+//            $this->error('修改失败请稍后重试');
+            return api_return(0, '修改失败请稍后重试');
         }
-        $this->success('修改成功');
+//        $this->success('修改成功');
+        return api_return(1, '修改成功');
     }
 
     // 删除购物车商品
@@ -193,24 +212,32 @@ class Cart extends ApiController
         $user = Member::where('id', $user_id)->find();
         if (empty($user)) {
             Sessions('member_id', null);
-            $this->error('用户信息异常请重新登录');
+//            $this->error('用户信息异常请重新登录');
+            return api_return(0, '用户信息异常请重新登录');
         }
         $post = $this->request->post();
         $cart_goods_ids = isset($post['cart_goods_ids']) ? $post['cart_goods_ids'] : null;
         if (!is_array($cart_goods_ids) && count($cart_goods_ids)) {
-            $this->error('传输参数有误或没有选中要删除的商品');
+//            $this->error('传输参数有误或没有选中要删除的商品');
+            return api_return(0, '传输参数有误或没有选中要删除的商品');
         }
         $cart_goods_objs = CartModel::where('uid', $user_id)->whereIn('id', implode(',', $cart_goods_ids))->select();
-        $cart_goods_objs->count() === 0 && $this->error('购物车商品不存在');
+//        $cart_goods_objs->count() === 0 && $this->error('购物车商品不存在');
+        if($cart_goods_objs->count()){
+            return api_return(0, '购物车商品不存在');
+        }
         try {
             $result = $cart_goods_objs->delete();
             if ($result === false) {
-                $this->error('删除失败请稍后重试');
+//                $this->error('删除失败请稍后重试');
+                return api_return(0, '删除失败请稍后重试');
             }
         } catch (Exception $e) {
-            $this->error('删除失败请稍后重试');
+//            $this->error('删除失败请稍后重试');
+            return api_return(0, '删除失败请稍后重试');
         }
-        $this->success('删除成功');
+//        $this->success('删除成功');
+        return api_return(1, '删除成功');
     }
